Move MemoryLabelRepositoryModule to apps/memory module
